refactor: move setScarecrow to its own file
