+++ src/game-logic/reducers/field.js
@@ -1,7 +1,6 @@
 import {
   cropLifeStage,
   fertilizerType,
-  fieldMode,
   itemType,
   toolLevel,
   toolType,
@@ -12,59 +11,22 @@ import {
   doesInventorySpaceRemain,
   getCropLifeStage,
   getInventoryQuantityMap,
-  getPlotContentFromItemId,
   getPlotContentType,
   getSeedItemIdFromFinalStageCropItemId,
   isRandomNumberLessThan,
 } from '../../utils'
-import {
-  HOE_LEVEL_TO_SEED_RECLAIM_RATE,
-  SCARECROW_ITEM_ID,
-} from '../../constants'
+import { HOE_LEVEL_TO_SEED_RECLAIM_RATE } from '../../constants'
 import { INVENTORY_FULL_NOTIFICATION } from '../../strings'
 import { ResourceFactory } from '../../factories'
 
 import { addItemToInventory } from './addItemToInventory'
-import { decrementItemFromInventory } from './decrementItemFromInventory'
 
 import { showNotification } from './showNotification'
 import { modifyFieldPlotAt } from './modifyFieldPlotAt'
 import { removeFieldPlotAt } from './removeFieldPlotAt'
 import { plantInPlot } from './plantInPlot'
 
-const { OBSERVE, SET_SCARECROW } = fieldMode
 const { GROWN } = cropLifeStage
-
-/**
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @returns {farmhand.state}
- */
-export const setScarecrow = (state, x, y) => {
-  const plot = state.field[y][x]
-
-  // Only set scarecrows in empty plots
-  if (plot !== null) {
-    return state
-  }
-
-  state = decrementItemFromInventory(state, SCARECROW_ITEM_ID)
-
-  const doScarecrowsRemain = state.inventory.some(
-    item => item.id === SCARECROW_ITEM_ID
-  )
-
-  state = modifyFieldPlotAt(state, x, y, () =>
-    getPlotContentFromItemId(SCARECROW_ITEM_ID)
-  )
-
-  return {
-    ...state,
-    fieldMode: doScarecrowsRemain ? SET_SCARECROW : OBSERVE,
-    selectedItemId: doScarecrowsRemain ? SCARECROW_ITEM_ID : '',
-  }
-}
 
 /**
  * @param {farmhand.state} state
