+++ src/game-logic/reducers/setScarecrow.js
@@ -1,0 +1,39 @@
+import { fieldMode } from '../../enums'
+import { getPlotContentFromItemId } from '../../utils'
+import { SCARECROW_ITEM_ID } from '../../constants'
+
+import { decrementItemFromInventory } from './decrementItemFromInventory'
+import { modifyFieldPlotAt } from './modifyFieldPlotAt'
+
+const { OBSERVE, SET_SCARECROW } = fieldMode
+
+/**
+ * @param {farmhand.state} state
+ * @param {number} x
+ * @param {number} y
+ * @returns {farmhand.state}
+ */
+export const setScarecrow = (state, x, y) => {
+  const plot = state.field[y][x]
+
+  // Only set scarecrows in empty plots
+  if (plot !== null) {
+    return state
+  }
+
+  state = decrementItemFromInventory(state, SCARECROW_ITEM_ID)
+
+  const doScarecrowsRemain = state.inventory.some(
+    item => item.id === SCARECROW_ITEM_ID
+  )
+
+  state = modifyFieldPlotAt(state, x, y, () =>
+    getPlotContentFromItemId(SCARECROW_ITEM_ID)
+  )
+
+  return {
+    ...state,
+    fieldMode: doScarecrowsRemain ? SET_SCARECROW : OBSERVE,
+    selectedItemId: doScarecrowsRemain ? SCARECROW_ITEM_ID : '',
+  }
+}
